Add private repost methods
